refactor: 도메인 별 분리 및 레이어드 아키텍처로 변경
