+++ src/main/java/com/chang/omg/domain/game/controller/dto/CharacterRankingResponse.java
@@ -1,0 +1,5 @@
+package com.chang.omg.domain.game.controller.dto;
+
+public record CharacterRankingResponse(String worldName, String characterName, Long count) {
+
+}
