+++ src/main/java/com/chang/omg/domain/game/service/api/kartrider/dto/User.java
@@ -1,0 +1,10 @@
+package com.chang.omg.domain.game.service.api.kartrider.dto;
+
+import java.util.List;
+
+public record User(List<OuidInfo> ouidInfo) {
+
+    public record OuidInfo(String ouid, String racerDateCreate, String racerLevel) {
+
+    }
+}
