+++ src/main/java/com/chang/omg/domain/game/service/api/kartrider/dto/UserBasic.java
@@ -1,0 +1,11 @@
+package com.chang.omg.domain.game.service.api.kartrider.dto;
+
+public record UserBasic(
+        String racerName,
+        String racerDateCreate,
+        String racerDateLastLogin,
+        String racerDateLastLogout,
+        String racerLevel
+) {
+
+}
