+++ src/main/java/com/chang/omg/domain/game/service/api/kartrider/dto/UserTitleEquipment.java
@@ -1,0 +1,10 @@
+package com.chang.omg.domain.game.service.api.kartrider.dto;
+
+import java.util.List;
+
+public record UserTitleEquipment(List<TitleEquipment> titleEquipment) {
+
+    record TitleEquipment(String titleName) {
+
+    }
+}
