+++ src/main/java/com/chang/omg/domain/game/service/api/maplestorym/dto/Character.java
@@ -1,0 +1,5 @@
+package com.chang.omg.domain.game.service.api.maplestorym.dto;
+
+public record Character(String ocid) {
+
+}
