+++ src/main/java/com/chang/omg/domain/game/service/api/maplestorym/dto/CharacterGuild.java
@@ -1,0 +1,5 @@
+package com.chang.omg.domain.game.service.api.maplestorym.dto;
+
+public record CharacterGuild(String guildName) {
+
+}
