+++ src/main/java/com/chang/omg/domain/game/service/api/maplestorym/dto/CharacterStat.java
@@ -1,0 +1,10 @@
+package com.chang.omg.domain.game.service.api.maplestorym.dto;
+
+import java.util.List;
+
+public record CharacterStat(List<Stat> stat) {
+
+    public record Stat(String statName, String statValue) {
+
+    }
+}
