+++ src/main/java/com/chang/omg/domain/rank/domain/CharacterInfo.java
@@ -1,0 +1,8 @@
+package com.chang.omg.domain.rank.domain;
+
+import lombok.Builder;
+
+@Builder
+public record CharacterInfo(String worldName, String characterName) {
+
+}
