--- src/main/java/com/chang/omg/domains/rank/domain/CharacterInfo.java
@@ -1,8 +1,0 @@
-package com.chang.omg.domains.rank.domain;
-
-import lombok.Builder;
-
-@Builder
-public record CharacterInfo(String worldName, String characterName) {
-
-}
