--- src/main/java/com/chang/omg/infrastructure/api/kartrider/KartRiderApi.java
@@ -1,16 +1,0 @@
-package com.chang.omg.infrastructure.api.kartrider;
-
-import com.chang.omg.infrastructure.api.kartrider.dto.User;
-import com.chang.omg.infrastructure.api.kartrider.dto.UserBasic;
-import com.chang.omg.infrastructure.api.kartrider.dto.UserTitleEquipment;
-
-import reactor.core.publisher.Mono;
-
-public interface KartRiderApi {
-
-    User getUserOuid(final String userName);
-
-    Mono<UserBasic> getUserBasic(final String ouid);
-
-    Mono<UserTitleEquipment> getUserTitleEquipment(final String ouid);
-}
