--- src/main/java/com/chang/omg/infrastructure/api/kartrider/dto/UserBasic.java
@@ -1,11 +1,0 @@
-package com.chang.omg.infrastructure.api.kartrider.dto;
-
-public record UserBasic(
-        String racerName,
-        String racerDateCreate,
-        String racerDateLastLogin,
-        String racerDateLastLogout,
-        String racerLevel
-) {
-
-}
