--- src/main/java/com/chang/omg/infrastructure/api/kartrider/dto/UserTitleEquipment.java
@@ -1,10 +1,0 @@
-package com.chang.omg.infrastructure.api.kartrider.dto;
-
-import java.util.List;
-
-public record UserTitleEquipment(List<TitleEquipment> titleEquipment) {
-
-    record TitleEquipment(String titleName) {
-
-    }
-}
