--- src/main/java/com/chang/omg/infrastructure/api/maplestorym/dto/Character.java
@@ -1,5 +1,0 @@
-package com.chang.omg.infrastructure.api.maplestorym.dto;
-
-public record Character(String ocid) {
-
-}
