--- src/main/java/com/chang/omg/infrastructure/api/maplestorym/dto/CharacterGuild.java
@@ -1,5 +1,0 @@
-package com.chang.omg.infrastructure.api.maplestorym.dto;
-
-public record CharacterGuild(String guildName) {
-
-}
