--- src/main/java/com/chang/omg/infrastructure/api/maplestorym/dto/CharacterItemEquipment.java
@@ -1,10 +1,0 @@
-package com.chang.omg.infrastructure.api.maplestorym.dto;
-
-import java.util.List;
-
-public record CharacterItemEquipment(List<ItemEquipment> itemEquipment) {
-
-    public record ItemEquipment(String itemName, String itemEquipmentPageName, String itemEquipmentSlotName) {
-
-    }
-}
