--- src/main/java/com/chang/omg/presentation/game/dto/CharacterRankingResponse.java
@@ -1,5 +1,0 @@
-package com.chang.omg.presentation.game.dto;
-
-public record CharacterRankingResponse(String worldName, String characterName, Long count) {
-
-}
